refactor: extract shared registry-version fetch into version.ts
diff.ts and check.ts each grew a near-identical npm-registry latest-version
fetch in the same session (#451, #452). Lift diff.ts's (correct) implementation
into version.ts as fetchRegistryLatestVersion + isComparableVersion so both
commands share one copy. Pure refactor for diff.ts; also renames the local
UpdateCache interface to UpdateCacheFile and comment-links the canonical shape.
