+++ packages/cli/src/commands/diff.ts
@@ -11,7 +11,11 @@ import { SAFEWORD_SCHEMA } from '../schema.js';
 import { createProjectContext } from '../utils/context.js';
 import { exists, readFileSafe, readJson } from '../utils/fs.js';
 import { error, header, info, listItem, success, warn } from '../utils/output.js';
-import { compareVersions } from '../utils/version.js';
+import {
+  compareVersions,
+  fetchRegistryLatestVersion,
+  isComparableVersion,
+} from '../utils/version.js';
 import { VERSION } from '../version.js';
 
 interface DiffOptions {
@@ -25,15 +29,13 @@ interface FileDiff {
   newContent?: string;
 }
 
-interface UpdateCache {
+/**
+ * Local view of `.update-cache.json` (canonical shape: templates/hooks/lib/update-cache.ts).
+ * `latestVersion` is `unknown` — it is untrusted cache data; validate via
+ * `isComparableVersion` before use.
+ */
+interface UpdateCacheFile {
   latestVersion?: unknown;
-}
-
-const REGISTRY_TIMEOUT_MS = 3000;
-const VERSION_PATTERN = /^\d+\.\d+\.\d+$/;
-
-function isComparableVersion(value: unknown): value is string {
-  return typeof value === 'string' && VERSION_PATTERN.test(value);
 }
 
 function newerOfCliAnd(candidate: string): string {
@@ -42,31 +44,8 @@ function newerOfCliAnd(candidate: string): string {
 
 function readCachedTargetVersion(safewordDirectory: string): string | undefined {
   const cachePath = nodePath.join(safewordDirectory, '.update-cache.json');
-  const cache = readJson(cachePath) as UpdateCache | undefined;
+  const cache = readJson(cachePath) as UpdateCacheFile | undefined;
   return isComparableVersion(cache?.latestVersion) ? cache.latestVersion : undefined;
-}
-
-async function fetchRegistryLatestVersion(
-  timeout = REGISTRY_TIMEOUT_MS,
-): Promise<string | undefined> {
-  const controller = new AbortController();
-  const timeoutId = setTimeout(() => {
-    controller.abort();
-  }, timeout);
-
-  try {
-    const response = await fetch('https://registry.npmjs.org/safeword/latest', {
-      signal: controller.signal,
-    });
-    if (!response.ok) return undefined;
-
-    const data = (await response.json()) as { version?: unknown };
-    return isComparableVersion(data.version) ? data.version : undefined;
-  } catch {
-    return undefined;
-  } finally {
-    clearTimeout(timeoutId);
-  }
 }
 
 async function resolveDiffTargetVersion(safewordDirectory: string): Promise<string> {
